Add new Route for payment

+++ controllers/adminController.js
@@ -1,4 +1,7 @@
 const pool = require("../config/db");
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
 
 // Get dashboard counts
 exports.getDashboardCounts = async (req, res) => {
@@ -109,12 +112,31 @@ exports.addDoctor = async (req, res) => {
     try {
         const { dr_name, dr_certificate, dr_position, dr_speciality, dr_contact, dr_email, dr_address, dr_gender, dr_experience, department_id, dr_fee, dr_about, dr_status } = req.body;
         const dr_photo = req.file ? req.file.filename : null;
+        const [existingUser] = await pool.execute("SELECT user_id FROM users WHERE user_email = ?", [dr_email]);
+
+        if (existingUser.length > 0) {
+            return res.status(400).json({ message: "Doctor email already has login account" });
+        }
 
         const [rows] = await pool.execute(
             `INSERT INTO doctors (dr_name, dr_certificate, dr_position, dr_speciality, dr_contact, dr_email, dr_photo, dr_address, dr_gender, dr_experience, department_id, dr_fee, dr_about, dr_status) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
             [dr_name, dr_certificate, dr_position, dr_speciality, dr_contact, dr_email, dr_photo, dr_address, dr_gender, dr_experience, department_id, dr_fee, dr_about, dr_status]
         );
-        res.json({ message: "Doctor added successfully", doctor_id: rows.insertId });
+
+        const cleanName = dr_name.replace(/^dr\.\s*/i, "").trim(); //remove Dr. from name
+        const firstName = cleanName.split(" ")[0];   // take first word
+        const plainPassword = `${firstName}@123`;
+        const hashedPassword = await bcrypt.hash(plainPassword, 10);
+
+        await pool.execute(
+            `INSERT INTO users (user_name, user_email, user_phone, user_gender, user_address, user_password, user_profile, role)VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
+            [dr_name, dr_email, dr_contact, dr_gender, dr_address, hashedPassword, dr_photo, "doctor"]);
+
+        res.json({
+            message: "Doctor added successfully",
+            doctor_id: rows.insertId,
+            tempPassword: plainPassword  // admin can see & share with doctor
+        });
     } catch (err) {
         console.error("addDoctor error", err);
         res.status(500).json({ error: err.message });
@@ -128,21 +150,35 @@ exports.updateDoctor = async (req, res) => {
         const { dr_name, dr_gender, dr_certificate, dr_position, dr_speciality, dr_contact, dr_email, dr_address, dr_experience, department_id, dr_fee, dr_about, dr_status } = req.body;
         const dr_photo = req.file ? req.file.filename : null;
 
+        //Get OLD email & image from doctors
+        const [oldDoctor] = await pool.execute("SELECT dr_email, dr_photo FROM doctors WHERE doctor_id = ?", [doctorId]);
+
+        if (oldDoctor.length === 0) {
+            return res.status(404).json({ message: "Doctor not found" });
+        }
+
+        const oldEmail = oldDoctor[0].dr_email;
+        const oldImage = oldDoctor[0].dr_photo;
+
         if (dr_photo === null) {
             await pool.execute(
                 `UPDATE doctors SET dr_name = ?, dr_gender = ?, dr_certificate = ?, dr_position = ?, dr_experience = ?, dr_speciality = ?, department_id = ?, dr_contact = ?, dr_fee = ?, dr_email = ?, dr_address = ?, dr_about = ?, dr_status = ? WHERE doctor_id = ?`,
                 [dr_name, dr_gender, dr_certificate, dr_position, dr_experience, dr_speciality, department_id, dr_contact, dr_fee, dr_email, dr_address, dr_about, dr_status, doctorId]
             );
-            res.json({ message: "Doctor updated successfully" });
-
         }
         else {
             await pool.execute(
                 `UPDATE doctors SET dr_name = ?, dr_gender = ?, dr_certificate = ?, dr_position = ?, dr_experience = ?, dr_speciality = ?, department_id = ?, dr_contact = ?, dr_fee = ?, dr_email = ?, dr_photo = ?, dr_address = ?, dr_about = ?, dr_status = ? WHERE doctor_id = ?`,
                 [dr_name, dr_gender, dr_certificate, dr_position, dr_experience, dr_speciality, department_id, dr_contact, dr_fee, dr_email, dr_photo, dr_address, dr_about, dr_status, doctorId]
             );
-            res.json({ message: "Doctor updated successfully" });
         }
+
+        await pool.execute(`UPDATE users SET user_name = ?, user_email = ?, user_phone = ?, user_gender = ?, user_address = ?, user_profile = ? WHERE user_email = ? AND role = 'doctor'`,
+            [dr_name, dr_email, dr_contact, dr_gender, dr_address, dr_photo ? dr_photo : oldImage, oldEmail]
+        ); //for if admin not update image put old image as it is use oldImage also.
+
+        res.json({ message: "Doctor updated successfully" });
+
     } catch (err) {
         console.error("Update doctor error", err);
         res.status(500).json({ error: err.message });
@@ -166,7 +202,13 @@ exports.getAllDoctors = async (req, res) => {
 exports.deleteDoctor = async (req, res) => {
     try {
         const doctorId = req.params.id;
+        const [rows] = await pool.execute("SELECT dr_email FROM doctors WHERE doctor_id = ?", [doctorId]);
+        if (rows.length === 0) {
+            return res.status(404).json({ message: "Doctor not found" });
+        }
+        const doctorEmail = rows[0].dr_email;
         await pool.execute("DELETE FROM doctors WHERE doctor_id = ?", [doctorId]);
+        await pool.execute("DELETE FROM users WHERE email = ? AND role = 'doctor'", [doctorEmail]);
         res.json({ message: "Doctor deleted successfully" });
     } catch (err) {
         console.error("deleteDoctor error", err);
